+++ src/pages/admin/orders/order-detail.tsx
@@ -1,0 +1,7 @@
+import OrderDetailPage from '@components/order/order-detail-page';
+
+const AdminOrderDetail = () => {
+  return <OrderDetailPage fallbackPath="/admin/orders" />;
+};
+
+export default AdminOrderDetail;
