+++ src/shared/components/order/order-list-table.tsx
@@ -29,6 +29,7 @@ export type OrderRow = {
 
 type Props = {
   data: OrderRow[];
+  detailBasePath?: string;
 };
 
 const STATUS_OPTIONS: Array<{ label: string; value: OrderStatusValue }> = [
@@ -140,7 +141,10 @@ const OrderStatusControl = ({ row }: { row: OrderRow }) => {
   );
 };
 
-const OrderListTable = ({ data }: Props) => {
+const OrderListTable = ({
+  data,
+  detailBasePath = '/app/orders',
+}: Props) => {
   const navigate = useNavigate();
   const [isHeaderFilterOpen, setIsHeaderFilterOpen] = useState(false);
   const [selectedStatus, setSelectedStatus] = useState<OrderStatusValue | 'ALL'>(
@@ -255,7 +259,7 @@ const OrderListTable = ({ data }: Props) => {
           {filteredData.map((row) => (
             <tr
               key={row.id}
-              onClick={() => navigate(`/app/orders/${row.orderId}`)}
+              onClick={() => navigate(`${detailBasePath}/${row.orderId}`)}
               className="h-[9.2rem] cursor-pointer border-b border-slate-200 transition hover:bg-[#FAFAFA] last:border-b-0"
             >
               <td className="px-2 text-center align-middle text-[1.5rem] font-normal text-slate-900 md:px-4 md:text-[1.8rem]">
